Trying to add a graphical uer interface
I tried to implement a section was added to the graphical user
interface of the Schrödinger equation solver, allowing users
to select various parameters, including physics parameters and solver options.
The updated interface should supports the modifications that I
made for a customization for both 1D and 2D simulations.